split api client functions into smaller files
