+++ pkg/internal/api/private_endpoints.go
@@ -1,0 +1,121 @@
+package api
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+	"strings"
+)
+
+type PrivateEndpoint struct {
+	CloudProvider string `json:"cloudProvider,omitempty"`
+	Description   string `json:"description,omitempty"`
+	EndpointId    string `json:"id,omitempty"`
+	Region        string `json:"region,omitempty"`
+}
+
+type OrgPrivateEndpointsUpdate struct {
+	Add    []PrivateEndpoint `json:"add,omitempty"`
+	Remove []PrivateEndpoint `json:"remove,omitempty"`
+}
+
+type OrganizationUpdate struct {
+	PrivateEndpoints *OrgPrivateEndpointsUpdate `json:"privateEndpoints"`
+}
+
+type OrgResult struct {
+	CreatedAt        string            `json:"createdAt,omitempty"`
+	ID               string            `json:"id,omitempty"`
+	Name             string            `json:"name,omitempty"`
+	PrivateEndpoints []PrivateEndpoint `json:"privateEndpoints,omitempty"`
+}
+
+type OrgPrivateEndpointConfig struct {
+	EndpointServiceId string `json:"endpointServiceId,omitempty"`
+}
+
+func (c *ClientImpl) GetServicePrivateEndpointConfig(ctx context.Context, serviceId string) (*ServicePrivateEndpointConfig, error) {
+	req, err := http.NewRequest(http.MethodGet, c.getServicePath(serviceId, "/privateEndpointConfig"), nil)
+	if err != nil {
+		return nil, err
+	}
+
+	body, err := c.doRequest(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+
+	endpointConfigResponse := ResponseWithResult[ServicePrivateEndpointConfig]{}
+	err = json.Unmarshal(body, &endpointConfigResponse)
+	if err != nil {
+		return nil, err
+	}
+
+	return &endpointConfigResponse.Result, nil
+}
+
+func (c *ClientImpl) GetOrgPrivateEndpointConfig(ctx context.Context, cloudProvider string, region string) (*OrgPrivateEndpointConfig, error) {
+	privateEndpointConfigPath := c.getPrivateEndpointConfigPath(cloudProvider, region)
+
+	req, err := http.NewRequest(http.MethodGet, privateEndpointConfigPath, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	body, err := c.doRequest(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+
+	privateEndpointConfigResponse := ResponseWithResult[OrgPrivateEndpointConfig]{}
+	if err = json.Unmarshal(body, &privateEndpointConfigResponse); err != nil {
+		return nil, err
+	}
+
+	return &privateEndpointConfigResponse.Result, nil
+}
+
+func (c *ClientImpl) GetOrganizationPrivateEndpoints(ctx context.Context) (*[]PrivateEndpoint, error) {
+	req, err := http.NewRequest(http.MethodGet, c.getOrgPath(""), nil)
+	if err != nil {
+		return nil, err
+	}
+
+	body, err := c.doRequest(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+
+	orgResponse := ResponseWithResult[OrgResult]{}
+	err = json.Unmarshal(body, &orgResponse)
+	if err != nil {
+		return nil, err
+	}
+
+	return &orgResponse.Result.PrivateEndpoints, nil
+}
+
+func (c *ClientImpl) UpdateOrganizationPrivateEndpoints(ctx context.Context, orgUpdate OrganizationUpdate) (*[]PrivateEndpoint, error) {
+	rb, err := json.Marshal(orgUpdate)
+	if err != nil {
+		return nil, err
+	}
+
+	req, err := http.NewRequest(http.MethodPatch, c.getOrgPath(""), strings.NewReader(string(rb)))
+	if err != nil {
+		return nil, err
+	}
+
+	body, err := c.doRequest(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+
+	orgResponse := ResponseWithResult[OrgResult]{}
+	err = json.Unmarshal(body, &orgResponse)
+	if err != nil {
+		return nil, err
+	}
+
+	return &orgResponse.Result.PrivateEndpoints, nil
+}
